+++ source/ProcessManager.Components/BusinessValidation/BusinessValidator.cs
@@ -12,8 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.ProcessManager.Abstractions.Components;
-using Energinet.DataHub.ProcessManager.Abstractions.Components.BusinessValidation;
+using Energinet.DataHub.ProcessManager.Components.Abstractions.BusinessValidation;
 using Microsoft.Extensions.Logging;
 
 namespace Energinet.DataHub.ProcessManager.Components.BusinessValidation;