+++ source/ProcessManager.Orchestrations.Abstractions/Processes/BRS_026_028/BRS_028/V1/Model/RequestCalculatedWholesaleServicesRejectedV1.cs
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.ProcessManager.Abstractions.Components.BusinessValidation;
-using Energinet.DataHub.ProcessManager.Components.ValueObjects;
+using Energinet.DataHub.ProcessManager.Components.Abstractions.BusinessValidation;
+using Energinet.DataHub.ProcessManager.Components.Abstractions.ValueObjects;
 
 namespace Energinet.DataHub.ProcessManager.Orchestrations.Abstractions.Processes.BRS_026_028.BRS_028.V1.Model;