+++ docs/ProcessManager.Orchestrations.Abstractions/ReleaseNotes/ReleaseNotes.md
@@ -1,11 +1,15 @@
 # ProcessManager.Orchestrations.Abstractions Release Notes
 
-## Version 0.21.0
+## Version 1.1.0
 
 - Add new columns to `OrchestrationInstance` and `StartOrchestrationInstanceV1`:
     - ActorMessageId
     - TransactionId
     - MeteringPointId
+
+## Version 1.0.0
+
+- Update dependent NuGet package.
 
 ## Version 0.20.0